+++ crates/teloxide-core/src/payloads/stop_message_live_location_inline.rs
@@ -2,7 +2,7 @@
 
 use serde::Serialize;
 
-use crate::types::{Message, ReplyMarkup};
+use crate::types::{ReplyMarkup, True};
 
 impl_payload! {
     /// Use this method to edit live location messages. A location can be edited until its live_period expires or editing is explicitly disabled by a call to [`StopMessageLiveLocation`]. On success, True is returned.
@@ -10,15 +10,11 @@ impl_payload! {
     /// See also: [`StopMessageLiveLocation`](crate::payloads::StopMessageLiveLocation)
     ///
     /// [`StopMessageLiveLocation`]: crate::payloads::StopMessageLiveLocation
-    #[derive(Debug, PartialEq, Clone, Serialize)]
-    pub StopMessageLiveLocationInline (StopMessageLiveLocationInlineSetters) => Message {
+    #[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize)]
+    pub StopMessageLiveLocationInline (StopMessageLiveLocationInlineSetters) => True {
         required {
             /// Identifier of the inline message
             pub inline_message_id: String [into],
-            /// Latitude of new location
-            pub latitude: f64,
-            /// Longitude of new location
-            pub longitude: f64,
         }
         optional {
             /// Additional interface options. A JSON-serialized object for an [inline keyboard], [custom reply keyboard], instructions to remove reply keyboard or to force a reply from the user.